+++ src/App.jsx
@@ -2,7 +2,7 @@ import { useState } from 'react'
 import reactLogo from './assets/react.svg'
 import viteLogo from '/vite.svg'
 import './App.css'
-import Home from './pages/Home';
+import Home from './pages/Home/Home';
 
 function App() {
   return <Home />;
